refactor(react-gui): put the shared hooks under hooks/, split by what they depend on
`lib/` was a name defined by negation ("not a feature, not h3-kit, not the
worker"), which is the same catch-all shape as the `hooks/` and `utils/`
directories this refactor is dissolving -- and it did not describe its own
contents: three of its five modules depend on nothing but React while
useLiveFetch talks to the CueMol worker, exactly like useCueMolEventListener
one directory over.

A hook now lives with whatever owns it, and the ones nothing owns are grouped
by dependency:

  hooks/react/   depends on React and nothing else -- useLatestRef,
                 useStaleGuard, useDebouncedCallback. ESLint rejects any
                 import of CueMol, IPC, @shared, @main or a feature from
                 here, so the name stays true instead of drifting.
  hooks/cuemol/  the React binding over worker/client -- useLiveFetch (its
                 siblings useCueMol / useCueMolEventListener / useMolEditCommit
                 follow in their own PR, which touches ~70 importers).

timing.ts is not a hook and moves to utils/. tritium/CLAUDE.md documents the
placement rule.

The .gitignore exception is widened to any application source: the Python
`lib/` rule swallowing a TypeScript directory is a trap for any future
`src/**/lib/`, not just the one it caught.
