+++ tritium/react-gui/src/renderer/hooks/useDensityMapPanel.ts
@@ -25,7 +25,7 @@ import {
     SEM_SCENE,
     SEM_ANY,
 } from '../event'
-import { useLiveFetch } from '@renderer/lib/useLiveFetch'
+import { useLiveFetch } from '@renderer/hooks/cuemol/useLiveFetch'
 
 export interface UseDensityMapPanelOptions {
     cm: AsyncCueMol | null
