+++ tritium/react-gui/src/renderer/hooks/useMolCoordObjects.ts
@@ -14,8 +14,8 @@ import type { AsyncCueMol } from '../worker/client/AsyncCueMol'
 import type { SceneObjectEntry } from '../worker/server/services/listSceneObjects.service'
 import { objectFilters } from '../h3-kit/ObjectSelect'
 import { SEM_OBJECT, SEM_ANY } from '../event'
-import { useLiveFetch } from '@renderer/lib/useLiveFetch'
-import { EVENT_BURST_DEBOUNCE_MS } from '@renderer/lib/timing'
+import { useLiveFetch } from '@renderer/hooks/cuemol/useLiveFetch'
+import { EVENT_BURST_DEBOUNCE_MS } from '@renderer/utils/timing'
 
 const EMPTY: SceneObjectEntry[] = []
