+++ tritium/react-gui/src/renderer/hooks/useMultiGradState.ts
@@ -15,8 +15,8 @@ import { useRef } from 'react'
 import type { AsyncCueMol } from '../worker/client/AsyncCueMol'
 import type { GetMultiGradStateResult } from '../worker/server/services/rendererColoring.service'
 import { SEM_OBJECT, SEM_RENDERER, SEM_ANY } from '../event'
-import { useLiveFetch } from '@renderer/lib/useLiveFetch'
-import { EVENT_BURST_DEBOUNCE_MS } from '@renderer/lib/timing'
+import { useLiveFetch } from '@renderer/hooks/cuemol/useLiveFetch'
+import { EVENT_BURST_DEBOUNCE_MS } from '@renderer/utils/timing'
 
 const EMPTY_STATE: GetMultiGradStateResult = {
     ok: false,
