+++ tritium/react-gui/src/renderer/hooks/usePaintCapableRenderers.ts
@@ -15,8 +15,8 @@ import { useRef } from 'react'
 import type { AsyncCueMol } from '../worker/client/AsyncCueMol'
 import type { PaintCapableRendererEntry } from '../worker/server/services/rendererColoring.service'
 import { SEM_OBJECT, SEM_RENDERER, SEM_SCENE, SEM_ANY } from '../event'
-import { useLiveFetch } from '@renderer/lib/useLiveFetch'
-import { EVENT_BURST_DEBOUNCE_MS } from '@renderer/lib/timing'
+import { useLiveFetch } from '@renderer/hooks/cuemol/useLiveFetch'
+import { EVENT_BURST_DEBOUNCE_MS } from '@renderer/utils/timing'
 
 // SEM_SCENE is included for the bulk-load path: a slow qsc load fires no
 // per-object ADDED events the pane can use (the first fetch resolves before
