+++ tritium/react-gui/src/renderer/hooks/usePaintColoringStyles.ts
@@ -16,8 +16,8 @@ import { useRef } from 'react'
 import type { AsyncCueMol } from '../worker/client/AsyncCueMol'
 import type { PaintColoringStyleEntry } from '../worker/server/services/rendererColoring.service'
 import { SEM_STYLE, SEM_ANY } from '../event'
-import { useLiveFetch } from '@renderer/lib/useLiveFetch'
-import { EVENT_BURST_DEBOUNCE_MS } from '@renderer/lib/timing'
+import { useLiveFetch } from '@renderer/hooks/cuemol/useLiveFetch'
+import { EVENT_BURST_DEBOUNCE_MS } from '@renderer/utils/timing'
 
 const EMPTY: PaintColoringStyleEntry[] = []
