+++ tritium/react-gui/src/renderer/hooks/useRendererColoringState.ts
@@ -16,8 +16,8 @@ import type {
     GetRendererColoringStateResult,
 } from '../worker/server/services/rendererColoring.service'
 import { SEM_OBJECT, SEM_RENDERER, SEM_SCENE, SEM_ANY } from '../event'
-import { useLiveFetch } from '@renderer/lib/useLiveFetch'
-import { EVENT_BURST_DEBOUNCE_MS } from '@renderer/lib/timing'
+import { useLiveFetch } from '@renderer/hooks/cuemol/useLiveFetch'
+import { EVENT_BURST_DEBOUNCE_MS } from '@renderer/utils/timing'
 
 // Listen for renderer events when editing a renderer's coloring and for
 // object events when editing an object's coloring. Combining the masks
