+++ tritium/react-gui/src/renderer/hooks/useTrajectory.ts
@@ -19,7 +19,7 @@ import type {
 } from '../worker/server/services/trajectory.service';
 import { SEM_OBJECT, SEM_ANY, SEM_CHANGED } from '../event';
 import { useCueMolEventListener } from './useCueMolEventListener';
-import { useStaleGuard } from '@renderer/lib/useStaleGuard';
+import { useStaleGuard } from '@renderer/hooks/react/useStaleGuard';
 
 interface UseTrajectoryOptions {
     cm: AsyncCueMol | null;
